project_appDev/requirements.txt
Flask
Flask-SQLAlchemy
email-validator
flask-security-too
bcrypt
seaborn
flask-restful
redis
celery[redis]
weasyprint==52.5
Flask-Caching
